Refactor(core): merged cpu and memory packages into core
