--- cpu/adm_accumulator.go
@@ -1,6 +1,0 @@
-package cpu
-
-// ACCUMULATOR addressing mode
-func (cpu CPU) admAccumulator() (uint8, uint8) {
-	return cpu.getARegister(), cpu.getBRegister()
-}
