--- cpu/op_carry.go
@@ -1,359 +1,0 @@
-package cpu
-
-import "github.com/snes-emu/gose/utils"
-
-// adc16 performs an add with carry 16bit operation the formula is: accumulator = accumulator + data + carry
-func (cpu *CPU) adc16(data uint16) uint16 {
-	var result uint16
-	if cpu.dFlag {
-		// Decimal mode on -> BCD arithmetic used
-		panic("TODO, d flag in adc needs to be implemented")
-		//result = (cpu.getCRegister() & 0x000f) + (data & 0x000f) + utils.BoolToUint16[cpu.cFlag] + (cpu.getCRegister() & 0x00f0) + (data & 0x00f0) + (cpu.C & 0x0f00) + (data & 0x0f00) + (cpu.getCRegister() & 0xf000) + (data & 0xf000)
-
-	} else {
-		// Decimal mode off -> binary arithmetic used
-		result = cpu.getCRegister() + data + utils.BoolToUint16[cpu.cFlag]
-		// Last bit value
-		cpu.nFlag = result&0x8000 != 0
-		// Signed artihmetic overflow
-		cpu.vFlag = (data^result)&^(data^cpu.getCRegister())&0x8000 != 0
-		cpu.zFlag = result == 0
-		// Unsigned carry
-		cpu.cFlag = (result < cpu.getCRegister()) || (result == cpu.getCRegister() && cpu.cFlag)
-
-	}
-
-	return result
-}
-
-// adc8 performs an add with carry 8bit operation the formula is: accumulator = accumulator + data + carry
-func (cpu *CPU) adc8(data uint8) uint8 {
-	var result uint8
-	if cpu.dFlag {
-		// Decimal mode on -> BCD arithmetic used
-		panic("TODO, d flag in adc needs to be implemented")
-	} else {
-		// Decimal mode off -> binary arithmetic used
-		result = cpu.getARegister() + data + utils.BoolToUint8[cpu.cFlag]
-		// Last bit value
-		cpu.nFlag = result&0x80 != 0
-		// Signed artihmetic overflow
-		cpu.vFlag = (data^result)&^(data^cpu.getARegister())&0x80 != 0
-		cpu.zFlag = result == 0
-		// Unsigned carry
-		cpu.cFlag = (result < cpu.getARegister()) || (result == cpu.getARegister() && cpu.cFlag)
-
-	}
-
-	return result
-}
-
-// sbc performs a substract with carry operation handling the 8/16 bit cases
-func (cpu *CPU) adc(dataLo, dataHi uint8) {
-	if cpu.mFlag {
-		cpu.setARegister(cpu.adc8(dataLo))
-	} else {
-		cpu.setCRegister(cpu.adc16(utils.JoinUint16(dataLo, dataHi)))
-	}
-}
-
-func (cpu *CPU) op61() {
-
-	dataLo, dataHi := cpu.admPDirectX()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op63() {
-
-	dataLo, dataHi := cpu.admStackS()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 5 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op65() {
-
-	dataLo, dataHi := cpu.admDirect()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 4 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op67() {
-
-	dataLo, dataHi := cpu.admBDirect()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op69() {
-
-	dataLo, dataHi := cpu.admImmediateM()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 3 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 3 - utils.BoolToUint16[cpu.mFlag]
-}
-
-func (cpu *CPU) op6D() {
-
-	dataLo, dataHi := cpu.admAbsolute()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 5 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 3
-}
-
-func (cpu *CPU) op6F() {
-
-	dataLo, dataHi := cpu.admLong()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 4
-}
-
-func (cpu *CPU) op71() {
-
-	dataLo, dataHi := cpu.admPDirectY()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0] - utils.BoolToUint16[cpu.xFlag]*utils.BoolToUint16[cpu.pFlag]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op72() {
-	dataLo, dataHi := cpu.admPDirect()
-
-	cpu.adc(dataLo, dataHi)
-
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op73() {
-
-	dataLo, dataHi := cpu.admPStackSY()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 8 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op75() {
-	dataLo, dataHi := cpu.admDirectX()
-
-	cpu.adc(dataLo, dataHi)
-
-	cpu.cycles += 5 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op77() {
-
-	dataLo, dataHi := cpu.admDirectX()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) op79() {
-	dataLo, dataHi := cpu.admAbsoluteX()
-
-	cpu.adc(dataLo, dataHi)
-
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.xFlag] + utils.BoolToUint16[cpu.mFlag]*utils.BoolToUint16[cpu.pFlag]
-	cpu.PC += 3
-}
-
-func (cpu *CPU) op7D() {
-
-	dataLo, dataHi := cpu.admAbsoluteY()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0] - utils.BoolToUint16[cpu.xFlag]*utils.BoolToUint16[cpu.pFlag]
-	cpu.PC += 3
-}
-
-func (cpu *CPU) op7F() {
-	dataLo, dataHi := cpu.admLongX()
-	cpu.adc(dataLo, dataHi)
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 4
-}
-
-// sbc16 performs a substract with carry 16bit operation the formula is: accumulator = accumulator - data - 1 + carry
-func (cpu *CPU) sbc16(data uint16) uint16 {
-	var result uint16
-	if cpu.dFlag {
-		// Decimal mode on -> BCD arithmetic used
-		panic("TODO, d flag in sbc needs to be implemented")
-
-	} else {
-		// Decimal mode off -> binary arithmetic used
-		result = cpu.getCRegister() - data - 1 + utils.BoolToUint16[cpu.cFlag]
-		// Last bit value
-		cpu.nFlag = result&0x8000 != 0
-		// Signed artihmetic overflow
-		cpu.vFlag = ((data+1)^cpu.getCRegister())&^((data+1)^result)&0x8000 != 0
-		cpu.zFlag = result == 0
-		// Unsigned carry
-		cpu.cFlag = cpu.getCRegister() >= data
-
-	}
-
-	return result
-}
-
-// sbc8 performs a substract with carry 8bit operation the formula is: accumulator = accumulator - data - 1 + carry
-func (cpu *CPU) sbc8(data uint8) uint8 {
-	var result uint8
-	if cpu.dFlag {
-		// Decimal mode on -> BCD arithmetic used
-		panic("TODO, d flag in sbc needs to be implemented")
-
-	} else {
-		// Decimal mode off -> binary arithmetic used
-		result = cpu.getARegister() - data - 1 + utils.BoolToUint8[cpu.cFlag]
-		// Last bit value
-		cpu.nFlag = result&0x80 != 0
-		// Signed artihmetic overflow
-		cpu.vFlag = ((data+1)^cpu.getARegister())&^((data+1)^result)&0x80 != 0
-		cpu.zFlag = result == 0
-		// Unsigned carry
-		cpu.cFlag = cpu.getARegister() >= data
-
-	}
-
-	return result
-}
-
-// sbc performs a substract with carry operation handling the 8/16 bit cases
-func (cpu *CPU) sbc(dataLo, dataHi uint8) {
-	if cpu.mFlag {
-		cpu.setARegister(cpu.sbc8(dataLo))
-	} else {
-		cpu.setCRegister(cpu.sbc16(utils.JoinUint16(dataLo, dataHi)))
-	}
-}
-
-func (cpu *CPU) opE1() {
-
-	dataLo, dataHi := cpu.admPDirectX()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opE3() {
-
-	dataLo, dataHi := cpu.admStackS()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 5 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opE5() {
-
-	dataLo, dataHi := cpu.admDirect()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 4 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opE7() {
-
-	dataLo, dataHi := cpu.admBDirect()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opE9() {
-
-	dataLo, dataHi := cpu.admImmediateM()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 3 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 3 - utils.BoolToUint16[cpu.mFlag]
-}
-
-func (cpu *CPU) opED() {
-
-	dataLo, dataHi := cpu.admAbsolute()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 5 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 3
-}
-
-func (cpu *CPU) opEF() {
-
-	dataLo, dataHi := cpu.admLong()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 4
-}
-
-func (cpu *CPU) opF1() {
-
-	dataLo, dataHi := cpu.admPDirectY()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0] - utils.BoolToUint16[cpu.xFlag]*utils.BoolToUint16[cpu.pFlag]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opF2() {
-	dataLo, dataHi := cpu.admPDirect()
-
-	cpu.sbc(dataLo, dataHi)
-
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opF3() {
-
-	dataLo, dataHi := cpu.admPStackSY()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 8 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opF5() {
-	dataLo, dataHi := cpu.admDirectX()
-
-	cpu.sbc(dataLo, dataHi)
-
-	cpu.cycles += 5 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opF7() {
-
-	dataLo, dataHi := cpu.admDirectX()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 7 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0]
-	cpu.PC += 2
-}
-
-func (cpu *CPU) opF9() {
-	dataLo, dataHi := cpu.admAbsoluteX()
-
-	cpu.sbc(dataLo, dataHi)
-
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.xFlag] + utils.BoolToUint16[cpu.mFlag]*utils.BoolToUint16[cpu.pFlag]
-	cpu.PC += 3
-}
-
-func (cpu *CPU) opFD() {
-
-	dataLo, dataHi := cpu.admAbsoluteY()
-	cpu.sbc(dataLo, dataHi)
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag] + utils.BoolToUint16[cpu.getDLRegister() == 0] - utils.BoolToUint16[cpu.xFlag]*utils.BoolToUint16[cpu.pFlag]
-	cpu.PC += 3
-}
-
-func (cpu *CPU) opFF() {
-	dataLo, dataHi := cpu.admLongX()
-
-	cpu.sbc(dataLo, dataHi)
-
-	cpu.cycles += 6 - utils.BoolToUint16[cpu.mFlag]
-	cpu.PC += 4
-}
