--- cpu/op_stop.go
@@ -1,1 +1,0 @@
-package cpu
